refactor(views): move change_password view to views.auth
