+++ django/cantusdb_project/main_app/views/auth.py
@@ -1,0 +1,18 @@
+from django.contrib.auth import update_session_auth_hash
+from django.contrib.auth.forms import PasswordChangeForm
+from django.contrib import messages
+from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+
+
+@login_required
+def change_password(request):
+    if request.method == "POST":
+        form = PasswordChangeForm(request.user, request.POST)
+        if form.is_valid():
+            user = form.save()
+            update_session_auth_hash(request, user)
+            messages.success(request, "Your password was successfully updated!")
+    else:
+        form = PasswordChangeForm(request.user)
+    return render(request, "registration/change_password.html", {"form": form})
